Accept more capture summary models
This line is due to the added model for String.valueOf(CharSequence)
| p;Joiner;false;add;(CharSequence);;Argument[0];Argument[this];taint;df-generated |

This line is added because `merge` calls `add` and the above line was added.
| p;Joiner;false;merge;(Joiner);;Argument[0];Argument[this];taint;df-generated |

This line is added because `FileOutputStream`'s constructor is now modeled as propagating taint, not just as a sink.
| p;PrivateFlowViaPublicInterface$SPI;true;openStream;();;Argument[this];ReturnValue;taint;df-generated |

This is due to the model for java.nio.file.Files.copy(Path,Path,CopyOption[]) (which seems wrong to me).
| p;Sinks;true;copyFileToDirectory;(Path,Path,CopyOption[]);;Argument[1].Element;ReturnValue;taint;df-generated |

This line is added because `java.net.URL.openStream()` is now modeled as propagating taint, not just as a sink.
| p;Sinks;true;readUrl;(URL,Charset);;Argument[0];ReturnValue;taint;df-generated |